Tweak code for speed.
Pull the dictionary_match case out from the loop.

There are potential savings to be had by spotting copycode being
the same for several successive len's, as this will mean copycode,
cmdcode and cost all don't need to be reevaluated within the loop.

This is probably only worth it for large len.